Introduced "datasets" (unstable)
You can now switch datasets in the settings.
On "Save settings" it will create the new folder for the dataset

- moved settingsauc.txt to mod root folder
- introduced setting "dataset" in settings and all its methods
- intoduced modfolder var in helpfunktions

known bugs
- dataset directory will only be created on save settings (not really a
bug)
- mod will crash when you try to load/save message from a dataset that
does not exist (which is a problem of the mod. it should have more
genereous error handling on load/save)
--> maybe introduce to create the directory on save messages when
directory does not exist (user did not click save in settings)
- load/save messages buttons don't grey out etc, as i am too lazy to do
the handling now

all in all it works if you know what you are doing